+++ variants/esp32s3/t5s3_epaper/platformio.ini
@@ -5,7 +5,7 @@ board_build.partition = default_16MB.csv
 board_check = true
 upload_protocol = esptool
 build_flags = -fno-strict-aliasing
-  ${esp32_base.build_flags} 
+  ${esp32s3_base.build_flags} 
   -I variants/esp32s3/t5s3_epaper
   -D T5_S3_EPAPER_PRO
   -D USE_EINK
